Extract the BundleStatus to own module
